made the requested chnages
